+++ src/backend/entities/Interest.ts
