+++ src/backend/entities/Investment.ts
