--- src/backend/model/Interest.ts
@@ -1,14 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, ManyToOne } from 'typeorm';
-import { User } from './User';
-
-@Entity()
-export class Interest {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column('text')
-  interest: string;
-
-  @ManyToOne(() => User, (user) => user.interests)
-  user: User;
-}
